[#712] Fix $ZTIMEOUT/$ZINTERRUPT issues if pop in direct mode and driven code has error
Both $ZTIMEOUT and $ZINTERRUPT have issues when their respective interrupts pop
while the process is in direct mode and the code fragment driven as a result gets
an error such as an undefined variable. Under these circumstances, for $ZTIMEOUT,
the ERRWZTIMEOUT message is not printed and for $ZINTERRUPT, a confusing mishmash
of messages (some duplicated) is output.

Both of these have the same root cause. YDB r1.24 had YDB#324 applied which fixed a
similar sort of error whose cause was improper identification of whether a counted
or direct mode frame was responsible for the error. The problem is that both the
ZTIMEOUT and ZINTERRUPT frames are counted but since they are transcendental code
frames not associated with programs, they aren't eligible to end the loop yet without
this fix they did.