[#725] Fix $GET(gvn) so it does not return incorrect values in rare cases on AARCH64/ARMV7L if db access method is MM
Background
----------
This is an issue that we noticed during in-house testing. A pre-existing `r124/ydb353` subtest failed
in one run on an ARMV7L box. The failure symptom was the following.

```diff
--- ydb353/ydb353.diff ---
26c26
< ^x(4)=5000
---
> ^x(4)=2370
31a32,33
> r124_0_395/ydb353/child_ydb353.mjo4
> TEST-E-FAIL : $get(^x(4)) : Actual = "237"_$C(3) : Expected = 2370
```

This subtest (written 2 years ago) runs 5 processes concurrently each of which update `^x(1)`,
`^x(2)`, `^x(3)`, `^x(4)`, `^x(5)` separately using an alternate sequence of $INCREMENT, $GET and
SET operations. For example, the node `^x(4)` goes through the values `1`, `1abcdefgh` (`$GET(gvn)`
and concatenate of `"abcdefgh"`), `2` (`$INCREMENT(gvn)`), `2abcdefgh`, and so on.

A `$GET(gvn)` is always attempted `after` a `$INCREMENT(gvn)` operation so the return value from
`$GET(gvn)` is always expected to be an increasing sequence of integers. In the failure case, we were
expecting the integer `2370` but instead got a string `234_$C(3)`.

After some analysis, came to the conclusion that when the `$GET(gvn)` operation started executing it saw
one state of the database block (that contains the node `^x(4)`) but before the `$GET(gvn)` operation
committed, the block stage had changed due to a concurrent update. This meant that the `$GET(gvn)`
ended up holding a garbage return value due to it examining an inconsistent view of the database
block. It should have restarted in this case since a block that it had read was changed concurrently
by another process. But it did not restart and this caused a corrupt/garbage value to be returned.

This test uses `$INCREMENT` and `NOISOLATION`. Interestingly `$INCREMENT` has a commit-time optimization
for transactions that are not in a `TSTART`/`TCOMMIT` fence (aka `non-TP`) and `NOISOLATION` has
a commit-time optimization for transactions that are in a `TSTART`/`TCOMMIT` fence (aka `TP`). I
initially suspected one of the two optimizations to have played a role in this failure.

But later test runs (thousands of runs) showed that this failure also happened in a stage of the test
where `NOISOLATION` was not turned `ON`. This meant the `NOISOLATION` related commit time optimization
is not related to this failure.

Enhancements to the test and later test runs (again thousands of runs) showed that the failure was
in the `$GET(gvn)` operation and not in `$INCREMENT(gvn)`. So the `$INCREMENT` related commit time
optimization is not related to this failure.

Later test runs showed that the failure happened only in the following conditions.

1) The database access method is MM.
2) The system is AARCH64 or ARMV7L with multiple cores.

Fix
---
* The issue was a missing memory barrier in between when the block header and the block contents are updated.
  In architectures where memory accesses can end up happening out of order, this can cause incorrect
  validation in the read-only fast path in `t_end.c` and result in incorrect return values from `$GET`.
  This is now fixed by a `MM_WRITE_MEMORY_BARRIER` call. A huge comment has been added to describe why
  only `MM` requires this fix and not `BG`.

Misc notes
----------
* From looking at the code, it seems like this issue should have existed in `YottaDB r1.30` and prior
  YottaDB releases on the `ARMV7L` and `AARCH64` too. But we have not seen any test failures till now
  in those older releases (`r1.24`, `r1.26`, `r1.28` or `r1.30`). The current suspicion is that we did
  not run enough tests to eventually demonstrate the issue in those releases. Towards that, I ran
  1000 tests using `r1.30` on 4 ARMV7L boxes and got 2 failures on one of the boxes. So the issue did exist
  in `r1.30` as expected. Just that it is very difficult to reproduce.