[YottaDB/DB/YDBTest#391] Adding common file extensions to copyright python script
**Mapping information**

The following require entries in the map.
```
   c start=" *" end="*"
   h start=" *" end="*"
   ctemplate start=" *" end="*"
   y start=" *" end="*"
   mpt start=";" end=";"
   m start=";" end=";"
   msg start="!" end="!"
   java start="*" end="*"
   rst start=".. #" end="#"
```

All the following will be taken care by the default case in code. So no entries required for these in the map.
```
    py start="#" end="#"
    yml start="#" end="#"
    toml start="#" end="#"
    pyi start="#" end="#"
    s start="#" end="#"
    csh start="#" end="#"
    si start="#" end="#"
    sh start="#" end="#"
    txt start="#" end="#"
    gtc start="#" end="#"
    awk start="#" end="#"
    exp start="#" end="#"
    awk start="#" end="#"
    sql start="#" end="#"
    in start="#" end="#"
    exp start="#" end="#"
    cmake start="#" end="#"
```

The following file though having `.cmake` extension doesn't have an ending character for copyright.
```
    git-watcher.cmake start="#"
```

**Misc**

`All rights reserved.` line was missed when trying to automatically add YottaDB copyright. With this change it is included as well.

**Note**

`needs_copyright.sh` in different repo's takes care of skipping files which doesn't need the copyright related changes so even if an extension doesn't apply to that repo it should be fine to have it in the mapping. Example case is `.txt` which is excluded in YDBPython where as it isn't in YDB.