[#737] $QUIT in M program invoked from ydb_ci() now works
`dollar_quit()` is the function responsible for implement `$QUIT`. And that did not take call-in
base frames into account. In the case of a `ydb_ci()` call, there is a call-in base frame that
is the parent of the user M code that gets invoked by the call-in.

In the non-call-in case, this function looks into the caller frame to determine whether the callee
returns a regular value or an alias value.

But if invoked through a call-in, an alias return value is impossible so one can skip this logic
entirely and return as it it is a regular return value.