[#739] MUPIP INTEG correctly handles NULL subscripts when no end key is specified
This commit fixes a bug in the upstream V6.3-010 change tested for by YDBTest#371. This change fixed a SIG-11 involving a GTM2STDNULLCOLL() when null subscripts were enabled with standard collation and mu_end_key was null because no end key was specified. However, as we were developing a test for this, we discovered that the MUPIP INTEG -subscript command that caused the SIG-11 was now causing a different SIG-11 when the STD2GTMNULLCOLL is done on mu_end_key as part of the cleanup to restore it to its prior state in the database. The fix involves checking if mu_end_key is NULL before attempting the STD2GTMNULLCOLL just as is done before attempting the GTM2STDNULLCOLL. In testing, we discovered that this bug is still present in upstream versions up to at least V6.3-014.