[YottaDB/Lang/YDBPython#2] Replaced spaces with tabs in #defines in libyottadb.h
This change was needed to simplify the extraction of constant names and values by YDBPython. Specifically, YDBPython#2 requires the programmatic generation of Python-level equivalents of Simple API constants to allow users of the wrapper to access these constant values in YDBPython applications.

These changes simplify the extraction of constant names by regularizing the whitespace formatting of constant definitions in `libyottadb.h`. Specifically, the previous intermixing of tabs and spaces between `#define` and the constant name necessitated the use of a general whitespace delimiter for breaking up these lines. However, this resulted in the breaking up of the values of the constants as well, e.g. `(0x7fffffffllu * 1000llu * 1000llu)` becomes 5 strings instead of 1. By correcting the space/tab intermingling, it becomes straightforward for YDBPython to then split the fields of the `#define` line by tabs, keeping space-delimited value strings whole and intact. This allows the constant extraction logic in YDBPython to be notably simplified.

Note that this is not needed in YDBRust, as that wrapper directly includes `libyottadb.h`. In contrast, `YDBGo` only exposes select constants from `libyottadb.h` as manually defined in `scripts/gen_error_codes.sh`.