[#654] Remove checks for libtinfo.so.5 on Ubuntu as YottaDB should now be built with libtinfo.so.6 on Ubuntu 20.04 LTS.
This commit removes the requirement for libtinfo.so.5 to be installed on Ubuntu. In Ubuntu 18.10, the libtinfo library was upgraded to version 6 which caused installs of older YottaDB versions built with the then-supported Ubuntu 18.04 LTS to fail with a %YDB-E-DLLNOOPEN because it couldn't find version 5 of libtinfo.so. Since the supported version of Ubuntu will be 20.04 LTS going forward, there is no longer a need to look for an old version of libtinfo.