[#744] Make %CDS^%H work for %DT>94657
This commit fixes a %CDS^%H bug that returned an empty string for %DAT when %DT was greater than 94657 (equivalent to a date February 28th, 2100). This returned an empty string because %CDS explicitly checked that %DT was less than 94658. The fix involved removing this check which allows 94568 to correctly return a date of March 1st, 2021 instead of an empty string. Values should work up to the maximum for $ZDATE which is called by %CDS^%H.