[#704] Fix YDBDISTUNVERIF error when YottaDB is invoked by valgrind
This commit fixes a %YDB-E-YDBDISTUNVERIF error when invoking YottaDB via valgrind. The problem is that valgrind changes the process_name to `MEMCHECK-AMD64-` or a platform-specific equivalent rather than `YOTTADB`. This causes the $ydb_dist verification to fail. The fix is to disable this check when YottaDB is invoked by valgrind (i.e. the global `process_name` starts with `MEMCHECK`).

This commit also fixes several valgrind uninitialized variable warnings in `valgrind MUPIP INTEG` that were caught by the test.