Update Debian Dockerfile to use Bullseye instead of Buster
This commit updates the Debian Dockerfile to use the newer Bullseye release instead of Buster. This is being done because the imptp-rs pipeline in YDBTest is failing due to an outdated version of cargo in Buster that is installing an outdated version of rust which doesn't support mem::take.

The go-pinning code is removed because it is no longer necessary since Bullseye's version of go is 1.15 which is newer than the minimum required version (1.13).

The bitnami/minideb image was also removed. While it reduced the file size from 1.41Gb to 1.36Gb, it had to be removed due to pipeline failures.