Fix icu errors when installing YDBPosix, YDBZlib and YDBCrypt plugins in UTF-8 mode
Internally, we discovered a couple issues with the ydbinstall script when installing YDBPosix, YDBZlib and YDBCrypt plugins together. Firstly, the YDBPosix install would unset $ydb_icu_version causing the installs to fail for YDBZlib and YDBCrypt. Secondly, the YDBZlib plugin would produce %YDB-E-ICUSYMNOTFOUND errors when compiling the M mode version of the plugin if the ICU version was specified as `default`. Both of these are fixed by removing unnecessary calls to ydb_env_set and ydb_env_unset from the YDBPosix install and by moving the command to compile YDBZlib in M mode to after the UTF-8 commands.

The command that failed was

`sudo ./ydbinstall --posix --zlib --encplugin --utf8 default --force-install`

This commit also makes a change to sr_unix/gtm_icu.c based on a review comment to YDB!998. This change was intended to be included in YDB!998 but was accidentally left out and this was not discovered until after YDB!998 had already been merged.