[#497] Remove ydbinstall requirement for --force-install to install a distribution built on the same system
This commit eliminates the need to use `--force-install` to install a distribution built on the same system. Previously, ydbinstall would require the `force-install` option to install YottaDB if the system was not supported even if it was installing a build of YottaDB that had just been built on the same system.

To make this change, there were 2 requirements. First, CMake now saves a file called `build_os_release` which is a copy of /etc/os-release for the system where it built YottaDB. Secondly, the ydbinstall script looks at this file and, if the OS and version match, it considers the system supported, removing the need for `--force-install`.