[#754] Add new ydbinstall script options to build and install YDB from a git repository
This commit adds the ``--from-source` and `--branch` options to the ydbinstall script to build and install YottaDB from a git repository.

The `--from-source` option does a git clone on the specified git repo, builds YottaDB from this repo and passes all relevant arguments on to the build's version of ydbinstall to install the build.

The `--branch` option specifies a git branch to build from when the `--from-source` option is also specified. This does a git checkout command to select the branch before building YottaDB. If the specified branch exists, it will build the branch. Otherwise, the build fails with an appropriate error message.