[#756] Remove ydb_env_set from ydbinstall option --zlib and fix ydb_env_set error message
The ydbinstall option `--zlib` sometimes produced a `ydbinstall: 29: [: UTF-8: unexpected operator` error message when running a `source ydb_env_set` which actually meant line 29 of ydb_env_set not line 29 of ydbinstall. After removing the `source ydb_env_set` call from the `--zlib` option, the zlib install still works fine so this commit removes the call to ydb_env_set from the `--zlib` option.

This commit also fixes the ydb_env_set error message which was due to the use of `==` instead of `=`. While `==` works fine in bash, the script uses sh which recognizes `=` inside of `[]`. This causes an error message because sh doesn't support `==`. The fix was to change the `==` on that line to `=`. The error message also showed up when running `source ydb_env_set` from a terminal on a system where `ydbinstall --zlib` produced this error message.