[#705] New --plugins-only ydbinstall option installs plugins for an existing YottaDB installation
This commit adds a new `--plugins-only` option to the ydbinstall script which allows for the YDBAIM, YDBCrypt, YDBOcto, YDBPosix and YDBZlib plugins to be installed for an existing YottaDB installation. Previously, the ydbinstall script could only install plugins for a new YottaDB installation.

The plugins selected via the options `--aim`, `--encplugin`, `--posix`, `-octo` and `--zlib` are installed to the YottaDB version pointed to by `--installdir` (if it exists). Otherwise, they are installed to `$ydb_dist` (if it exists) or to the version pointed to by pkg-config if neither --installdir or `$ydb_dist` is set. If no plugins are selected, the `--plugins-only` option will do nothing.