Fix cmake minimum version error for --posix ydbinstall option on RHEL7
This commit fixes an issue with the `--posix` option on RHEL7 that was caused by the `--posix` option calling `cmake` without checking if the OS was Red Hat first. Since the cmake command on Red Hat is supposed to be `cmake3` the `--posix` option called an old version of cmake on an in-house machine causing the sudo/plugins and sudo/pluginsonly tests to fail. The fix is to call `cmake3` on machines running RHEL as we do everywhere else in the ydbinstall script that calls cmake.

This commit also changes the `git clone` in the `--aim` option to run with the `-q` option which suppresses the `Cloning into '.'...` message. This message caused sudo/plugins and sudo/pluginsonly test failures due to its absence on RHEL7 and is not important for the YDBAIM install.