Fix regression in 8c554336 that caused socket file descriptors to be inherited across an exec() in PRO builds
The commit 8c5543360 changed the build such that PRO builds no longer used the wrappers that
YDB uses for close/open/socket/etc  calls to implement file descriptor tracing. In order for
sockets to be created with the O_CLOEXEC flag (closes the socket in a fork'd process when
that process does an execve()), we need to ALWAYS setup gtm_socket() instead of just when
GTM_FD_TRACE is defined and change gtm_socket.h so the define of socket() to gtm_socket()
ALWAYS happens. Failure to do this causes the v63000/gtm8009 test to fail because it checks
that no filedescriptors survive in the child fork done by the ZSYSTEM command. What showed
up were sockets when it used a R/O database (socket created to communicate with gtmsecshr)
and when it intentionally tested a socket. This issue is only with the PRO build and is not
(yet) in any released software so no user interaction at this point.