Remove unnecessary sudo usages from ydbinstall
This commit removes 15 unnecessary usages of `sudo` from the ydbinstall script. These usages are unneeded because the ydbinstall script is usually run with root privileges. If ydbinstall is not run with root privileges, a `--from-source` or a plugin install should also not be run with root privileges. When `sudo` does not exist, these usages caused ydbinstall to fail with an error.