ydbinstall.sh script returns non-zero status when plugin install fails due to download error
The sudo/plugins test failed in weekend tests with a curl error from the `--zlib` option. While this curl failure was not itself an issue, the ydbinstall script returned a status of 0 for successful install when the zlib install had failed and the curl error only appeared in the test output because it was sent to stderr which isn't ignored by the test even if the install is successful.  This commit ensures that ydbinstall will correctly return a non-zero status when downloading a plugin fails.