[#785] ydb_ci_t() and ydb_cip_t() supply error text in errstr for errors occuring in M code
Previously, no text message was provided which made errors like LVUNDEF of limited use since
one did not know what variable was called out. The same thing happens with ANY unhandled error
that occurs in M code.

With this change, if such an error occurs, the errstr is filled in with the current value of
$ZSTATUS before releasing the engine lock.