[#793] Fix Bug with Comments at Beginning of External Call Tables
This commit fixes a bug with comments at the beginning of an external call table. If a comment appeared in the external call table before the shareable library name, the comment line would be interpreted as the name of the shareable library causing all external calls to not work. This change has been tested with comment lines prior to the shareable library name and comments on the same line after a shareable library name. Both now work as expected with the changes.