Add support Home and Keys keys for Direct Mode and READ mode.
For console and terminal (like VT100) modes.

I have added keycodes for terminal and console modes. In the latter case, via escape sequences.
And also moved the repeated sections of the code into macros.

The code for handling Home and End buttons I copied the existing one for
Ctrl-A (cursor to start of line )and Ctrl-E (to end of line).

Signed-off-by: Sam Habiel <sam@yottadb.com>