[#758] New %PEEKBYNAME() option node_local.max_procs reports maximum number of processes concurrently accessing a database file
This commit adds 2 new fields to the database headers to provide a %PEEKBYNAME() option to measure the maximum number of processes accessing a database.

* The `sgmnt_data.max_procs` field is intended as an internal field to store the largest count of concurrent processes accessing the database and the timestamp. It is placed inside of `sgmnt_data` to ensure that the values persist even after database load falls to 0.
* The `node_local.max_procs` field is a user-facing field that reports the fields of `sgmnt_data.max_procs` in the following format: `cnt,time` where the `cnt` variable is a process count and the `time` variables is timestamp in Unix time (seconds since the Unix epoch started in 1970).

When the contents of `node_local.ref_cnt` are incremented (because a new process is accessing the database), the contents of those fields are automatically updated via a macro if the number of processes accessing the database is larger than the previous maximum.

This commit also adds the new values to `DSE DUMP -FILEHEADER` and adds a new `DSE CHANGE -FILEHEADER -RESET_MAX_PROCS` option that resets the maximum process count to 0.

Since the database headers have been changed, this commit introduces a new minor database version and ensures that the fields for `sgmnt_data.max_procs` are initialized to 0.