[VALGRIND] Fix uninitialized memory in trigger_update_rec (trigger_update.c:1617)
Memory issue shown by Valgrind.

```
==98521== Conditional jump or move depends on uninitialised value(s)
==98521==    at 0x4F8DDEE: trigger_update_rec (trigger_update.c:1617)
==98521==    by 0x4EA0A26: trigger_update_rec_helper.constprop.61 (trigger_update.c:2162)
==98521==    by 0x4F1AF12: UnknownInlinedFun (trigger_update.c:2215)
==98521==    by 0x4F1AF12: op_fnztrigger (op_fnztrigger.c:248)
==98521==    by 0xA550DB4: _ydboctoplanhelpers (in /builds/shabiel/YDBOcto/build/src/utf8/_ydbocto.so)
==98521==    by 0x76E29EF: ???
==98521==    by 0x1FFEFFAB2F: ???
==98521==    by 0x100000001: ???
==98521==  Uninitialised value was created by a stack allocation
==98521==    at 0x4F8BC35: trigger_update_rec (trigger_update.c:1148)
==98521==
```

@nars1 did the analysis; copying and pasting his analysis:

One would think this is a false positive. This is because even though
reg_index and max_reg_index are not initialized in the else block at
line 1584 below, when they are used in 1617 they are guaranteed to be
initialized because if we had gone through the else block, gvspan is
guaranteed to be NULL and in that case we would have already done a
break; out of the loop at line 1616.

** sr_unix/trigger_update.c **

```c
   1562         if (NULL != gvspan)
   1563         {
   1564                 gvnh_spanreg_subs_gvt_init(gvnh_reg, gd_header,
			NULL);
   1565                 reg_index = gvspan->min_reg_index;
   1566                 min_reg_index = gvspan->min_reg_index;
   1567                 max_reg_index = gvspan->max_reg_index;
      .
      .
   1583         } else
   1584         {
   1585                 memcpy(disp_trigvn, trigvn, trigvn_len);
   1586                 disp_trigvn_len = trigvn_len;
   1587                 disp_trigvn[disp_trigvn_len] = '\0';    /* null
			terminate just in case */
   1588         }
      .
      .
   1615                 if (NULL == gvspan)
   1616                         break;
   1617                 if (reg_index >= max_reg_index)
   1618                         break;
```

But with instruction reordering optimizations happening in the compiler,
it is mostly the case that lines 1617-1618 got moved to BEFORE lines
1615-1616. This is valid because the variables used in those 2 if
conditions don't seem to be related and so can be freely reordered.
Given the compiler did that reordering, we now have an uninitialized
memory usage.

So the right fix is to initialize reg_index etc. to 0 in line 1585