[#829] Fix ydbinstall.sh for Octo on CentOS 7/RHEL 7
The error from `./ydbinstall.sh` reported by the user is as follows:
```
 Now installing YDBOcto
Cloning into 'YDBOcto-master'...
error: RPC failed; result=22, HTTP code = 422
fatal: The remote end hung up unexpectedly
Unable to download YDBOcto. Your internet connection and/or the gitlab servers may be down. Please try again later.
```

The issue is outlined here:
https://stackoverflow.com/questions/32533379/git-error-rpc-failed-result-22-http-code-404
and the accompanying Gitlab issue
https://gitlab.com/gitlab-org/gitlab/-/issues/29629.

The issue is on Centos 7, the old git version does something which makes
it not work on a URL that doesn't have .git on the end, that looks like
this: https://gitlab.com/YottaDB/DBMS/YDBOcto. On all other systems,
it's not a problem. In the same ydbinstall.sh file, YDBAIM is there with
.git, and that succeeds.

Commit adds .git to YDBOcto github URL.