[YDB#602] [V63011] Add alternate signal handling for SIGHUP now that it is catchable
Presently, only Go main routines calling into YottaDB utilize alternate signal handling where signal handling
is done by Go who then notifies YottaDB that signal(s) have occurred. So a Go main making calls into YottaDB
via the Go wrapper can also catch SIGHUP though YottaDB will treat them as a process-ending signal because
standard SIGHUP support in YottaDB is to drive the ^C handler but in call-in mode, ^C is treated like SIGTERM.
To avoid SIGTERM-like handling on a SIGHUP, the user will need to use the new user signal handler feature added with
YottaDB/Lang/YDBGo#34 and choose to NOT drive the YottaDB handler.

To explain a bit more, when a SIGHUP is encountered by YottaDB by a non-Go call-in initializes YottaDB, the signal
handler the ^C drives is generic_signal_handler() instead of ctrlc_handler(). The change in this issue for alternate
signal handling is doing much the same. A SIGHUP will be handled (mostly) as if it were a ^C. This change adds the
alternate handler for SIGHUP to be generic_signal_handler() and morphs the signal to a SIGINT in
ydb_altmain_sighandler(). Note that in generic_signal_handler(), SIGTERM and SIGINT are treated identically - both
stop the process but both are deferrable.