Fix incorrect assert failure in online_bkup_1/online6 test
The online_bkup_1/online6 test failed with a core in internal testing because it attempted to connect to a socket on a host that was unreachable, returning an `EHOSTUNREACH` error and the code in sr_port/tcp_open.c considered `ECONNREFUSED` the only acceptable error. This caused the test to fail and dump a core.

This particular test failure was only possible on debug builds as `assert` is only executed on debug builds. This commit fixes the failure by changing the assert(FALSE) to assert that the error is an `EHOSTUNREACH` which ensures that the function `tcp_open()` will now return a `-1` status on debug builds when the host is unreachable just as it does on production builds.