[#837] Remove unnecessary `sudo` usage in ydbinstall.sh's `--encplugin` option
This commit removes an unnecessary `sudo` usage in the `--encplugin` option of `ydbinstall.sh` that installs the YDBEncrypt plugin. Since the `ydbinstall.sh` script has to be run as root anyway, this `sudo` did nothing on machines where `sudo` is installed and would cause an error if `sudo` was not installed.