[#836] Change ydb_call_variadic_plist() parm from intptr_t to gparam_list*
This change was primarily needed for the Python wrapper. The new type is the actual type of the parameter.
It was earlier being disguised as an intptr due either to the version of Go being used when it was initially
written a few years back and/or due to the lack of Go experience by the author. Regardless of why, this flaw
is being corrected now making the types match for a cleaner interface.

This change necessitated the gparam_list structure moving from where it was defined in callg.h to a
separate header file along with some definitions that were in mdef.h that define the size of the argument
list in the structure. The definitions in callg.h and mdef.h respectively were replaced with an include of
the new header file to get those definitions. The reason for doing this is because when a user program
includes libyottadb.h, mdef.h does not get pulled in (and is not available to users generally). But now,
libyottadb.h also includes gparam_list.h so this structure is defined and available within the wrappers.

In addition, the new gparam_list.h include is now also installed in $ydb_dist by the install procedure due to
an update to CMakeLists.txt.

Lastly, the signature for function called by ydb_call_variadic_plist_func() was changed to remove a required
"count-of-parameters" value as the first argument of a variadic list. While earl versions of this call required
the count parameter, this was removed later but not from the function type's parameter list.