[#766] Changes to ydbinstall and configure to fix gnome crash on Ubuntu 20.04 caused by restarting systemd-logind in configure
YottaDB may not function correctly if `RemoveIPC=yes` is configured for systemd because systemd will remove IPCs for a non-system user's process when that user logs out. This can cause database operations to fail with mysterious errors. To address this, the `ydbinstall` script sets `RemoveIPC=no` for systemd unless the `--preserveRemoveIPC` command line option is used.

The problem is that YottaDB, by default, would restart `systemd-logind` in the `configure` script if `RemoveIPC` was set to `yes`. This was usually fine but it caused gnome to crash on Ubuntu 20.04. This commit removes the automatic restart of `systemd-logind` and replaces it with a message informing users who had `RemoveIPC` set to `yes` that YottaDB may function incorrectly restart `systemd-logind` or reboot their computer. Previously, `ydbinstall` would abort if the user selected `--preserveRemoveIPC` and `RemoveIPC` was set to `yes`. This has been changed to inform the user that YottaDB may not function correctly until they change `RemoveIPC` to `no` and restart either `systemd-logind` or their computer.