[#783] Set `$ZROUTINES` to `$ydb_dist/utf8/libyottadbutil.so` if `ydb_chset=UTF-8` and `ydb_routines` is not set
Previously, if `$ydb_routines` was not set, `$ZROUTINES` got set to
$ydb_dist/libyottadbutil.so or $ydb_dist if libyottadbutil.so was not
installed. This is incorrect in UTF-8 mode, as we need to use the
objects compiled in UTF-8 mode; thus $ydb_dist/utf8/libyottadbutil.so
or $ydb_dist/utf8/.

In addition, this change issues the error message UTF8NOTINSTALLED
(`$ydb_dist does not have utf8 folder installed. Please use M mode or
re-install YottaDB with UTF-8 support`) if YottaDB is installed without
UTF-8 support but is still invoked with `ydb_chset=UTF-8`.

In an unrelated change, per @estess, remove `!/ansi=0` from a couple of
other error messages as that code isn't used.