[#849] Enhanced Dockerfile-debian to have YDBPython installed and ready to use
Now that YDBPython is production ready, it may be useful for users to test it out using a YDB Docker image. To this, end YDBPython is now installed and tested by default in `Dockerfile-debian`.

This change entailed adding a `git clone` of the latest YDBPython master, as well as installing several pre-requisite packages via `apt`.