[#603] [V63012] Address clang-tidy pipeline failure
This commit addresses the clang-tidy pipeline failures by updating the clang-tidy warning files to reflect
V6.3-012 changes and by fixing a clang-tidy warning in `gtm_putmsg.list`. The change to `gtm_putmsg_list.c`
fixes the following warning:

`gtm_putmsg_list.c:warning: Value stored to 'f_actual' is never read [clang-analyzer-deadcode.DeadStores]`

As the `f_actual` store is never used, the right fix is to not set it and only set `f_count` instead.
This addresses the warning. The other changes in this commit are to adjust the amount of padding for 2
warnings in `gtm_threadgbl_deftypes` and remove a couple warnings on pro builds that are fixed by V6.3-012.
These changes should fix the clang-tidy pipeline.