[YottaDB/Lang/YDBGo#34] Move ERR_STRUCTNOTALLOCD and ERR_INVLNPAIRLIST out of ydberrors.msg
These are Go-only errors messages so are in-process of moving over to a new error facility in
YDBGo. This initial change just permits these errors to live in a new (manually created) copy
of error_local.go which puts the errors in a Go structure for use there but in future versions
will be auto-generated from YDBGo/ydbgoerrors.msg.