[#603] [V63012] Fix v53003/D9H12002672 test failure by restoring if (mupip_debug) block in sr_unix/ss_initiate.c
With the V6.3-012 changes merged into YottaDB, the v53003/D9H12002672 test would fail occasionally due to an extra
`Start kill-in-prog wait` message that appeared in its `MUPIP INTEG`. This message would show up whenever the test
system randomly chose an online integ. This was caused by the removal of an `if (debug_mupip)` around a `GET_CUR_TIME`
and a `util_out_print` in `sr_unix/ss_initiate.c`. This would print an extra message. Suspiciously, a later
`if (debug_mupip)` block that printed `Done with kill-in-prog wait` was left inside the `if`. As this upstream change
appears to be incorrect, the fix is to restore the `if (debug_mupip)`. With this restored, the test now passes reliably.

This block in `sr_unix/ss_initiate.c` has very similar code to `sr_port/mupip_backup.c` and `sr_port/region_freeze.c`
so reverting this is fine to maintain consistency with other code controlled by `debug_mupip`.