[#603] [V63012] Fix r130/ydb607 test failure caused by incorrect merge conflict resolution in sr_port/db_auto_upgrade.c
This commit fixes a r130/ydb607 test failure caused by an incorrect merge conflict resolution in
`sr_port/db_auto_upgrade.c`. The test selects an old version of YottaDB or GT.M between V6.3-003 and V6.3-006 and
YottaDB versions based on those GT.M versions. Whenever the test selected YottaDB versions r1.22 or r1.24, it would
fail due to the return of the YDB#607 bug. The cause was an incorrect `break` statement in the minor version for
V6.3-012 which caused the upgrade for versions with the minor version for r1.22 (which is the same as V6.3-007's)
to break out of the loop early and not do the YottaDB updates to the database including the update that fixed the
bug. The fix is to remove this `break` and an additional incorrect `break` after the block for r1.34's minor version.