[#603] [V63012] Fix v54002/C9B03001660 test failure by adding missed REVERT in sr_port/m_for.c
After the changes to fix the `dqnoop()` clang compiler warning in
bc477fa8bedbb6c634c20877260d295a5660b536, the v54002/C9B03001660 test started
failing with SIG-11s due to a missed REVERT in `sr_port/m_for.c`. This failure
can be illustrated with the following 2 lines of M code:

```
 xecute "for i=1"
 for a=1:1:2 kill a
```

The `dqnoop()` changes exposed this issue which is a merge conflict caused
by YottaDB having an `ESTABLISH` in this file (and thus needing a `REVERT`)
for every return path but GT.M not having an `ESTABLISH` in the file. By
adding the revert, this fixes the failure. A `v63012/missedrevert` test will
be added using the above 2 lines of M to test for this bug.