[#864] Fix V6.3-012 change to MUPIP BACKUP so that file paths > 255 characters return a FILENAMETOOLONG error
This commit fixes an issue related to the GTM-9182 changes from V6.3-012 that was discovered while
trying to write a test for the issue.

The GTM-9182 release note claims that `MUPIP BACKUP` issues a `FILENAMETOOLONG` error if the
backup file path exceeds the 255 character limit. A prior change in V6.3-003 (GTM-4212) added a
`FILENAMETOOLONG` error if the file path + the 24 character temporary file name exceeded the limit
so this change only applies if the database name is longer than the temporary file path. Since this
includes a leading `/`, backing up a filename of `mumpsmumpsmumpsmumps.dat` to a directory whose
absolute path length is 230 characters should cause an error. However, in testing (for YDBTest#419),
we found that this error did not happen with a file name that was `mumps` repeated around 30 times
except when it triggered the GTM-4212 code. This incorrect behavior happened whether the directory
path was specified as an absolute or relative path.

Upon examining the code change related to this change (which is clearly labeled as `GTM-9182` in
a comment), we discovered that the if block was never executed if a backup directory path was
passed. Specifically, the for loop would set `index` to the first `/`. Then the if block checked
if `index` was `>= file->len` which was impossible because `MUPIP BACKUP` always appended
`/mumps.dat` (or whatever the database file was called) at the end of the argument passed to the
command. Because of this, the first comparison was always false and the if block that would return
a `FILENAMETOOLONG` was always skipped. However, if a relative or absolute path to a file name was
passed where the full backup file path exceeded the 255 character limit, upstream version V6.3-012
would produce a `FILENAMETOOLONG` error.

This commit fixes the GTM-9182 code change by checking if file->len is greater than `MAX_FN_LEN`,
which is currently 255. If it is, it returns a FILENAMETOOLONG error. After examining the issue,
we decided that a relative path should be fine if it never overflows a buffer even if the absolute
path would be over the 255 character limit. Therefore, some relative file names that produce errors
on upstream will work fine with our change but our change also handles cases where MUPIP BACKUP is
passed a path of over 255 characters that ends with a directory name or where `ydb_baktmpdir` is
set to something that is less than the 255 character limit. In testing with this change, it behaves
exactly as expected. It has also reliably passed D_ALLs on both dbg and pro builds.

This commit also fixes an issue with the GTM-4212 changes from V6.3-003 and the GTM-9182 changes
where a `MUPIP BACKUP -noonline` with a file name or path > 255 characters would overflow the buffer
causing ASAN builds to crash with errors. The issue is that the code does not do any checks if the
file name/path is too long. Both the GTM-4212 and GTM-9182 checks were within an `if (TRUE == online)`
block but the `else` block that is invoked on a `-noonline` backup lacked the necessary checks. After
investigating, it was possible to do the equivalent of the new YDB#864 check in the `else` block but
it was impossible to do the GTM-4212 check within the `mupip_backup()` function. Instead, it has to
be done within `sr_port/mubfilcpy.c.`. Once both fixes are applied, the appropriate error messages
show up for `-noonline` backups.

This commit also makes a change to fix a duplicate `FILENAMETOOLONG` error. The duplicate appeared
because a `gtm_file_stat()` call produced a `FILENAMETOOLONG`. Then, `MUPIP BACKUP` issued a
`FILEPARSE` error and passed `ustatus` to `mupip_exit()` which produced a `FILENAMETOOLONG` error
message. The fix for the duplicate error is to pass `0` to `mupip_exit()` which ensures that it
does not produce duplicate `FILENAMETOOLONG` error messages. The `FILEPARSE` error in this block
also returns the full path to the directory. This required a call to `GET_CWD()` to return the rest
of the full path in `tempdir_full` so that it can be reconstructed before returning the `FILEPARSE`
error.

This commit also changes some `memcmp` calls to `STRNCMP_STR` calls because the `memcmp` calls
were overflowing the buffer for the string literal passed as the second argument which caused
ASAN failures. This fixes the ASAN failures.